Factoring our calibration code
